requirements.txt
streamlit
google-generativeai
python-dotenv
fsspec[http]<=2024.2.0,>=2023.1.0